fix: cast flags to uint for SetWindowPos
SetWindowPos expects uFlags as uint, but the flags variable was
inferred as int from the bitwise OR operations on untyped constants.
This fixes the Windows build error.

+++ v3/pkg/application/webview_window_windows.go
@@ -1597,7 +1597,7 @@ func (w *windowsWebviewWindow) WndProc(msg uint32, wparam, lparam uintptr) uintp
 				int(newWindowRect.Top),
 				int(newWindowRect.Right-newWindowRect.Left),
 				int(newWindowRect.Bottom-newWindowRect.Top),
-				flags)
+				uint(flags))
 			// For frameless windows with decorations, re-extend the frame into client area
 			// to ensure proper window frame styling after DPI change.
 			if w.framelessWithDecorations() {
